+++ ntex-tls/CHANGES.md
@@ -1,6 +1,6 @@
 # Changes
 
-## [4.0.0-beta.0] - 2026-08-24
+## [4.0.0] - 2026-09-14
 
 * Migrate to ntex-service 5